+++ exporter/opentelemetry-exporter-otlp-json-file/tests/test_log_exporter.py
@@ -7,9 +7,8 @@ import io
 import json
 import logging
 import unittest
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
 
-import opentelemetry.exporter.otlp.json.file._log_exporter as _log_exporter_mod
 from opentelemetry._logs import LogRecord, SeverityNumber
 from opentelemetry.exporter.otlp.json.common._log_encoder import encode_logs
 from opentelemetry.exporter.otlp.json.file._internal import _format_line
@@ -28,6 +27,8 @@ from opentelemetry.sdk._logs.export import (
 )
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.util.instrumentation import InstrumentationScope
+
+_LOGGER_NAME = "opentelemetry.exporter.otlp.json.file._log_exporter"
 
 
 def _make_log_record(
@@ -55,21 +56,15 @@ class TestFileLogExporter(unittest.TestCase):
         self.assertEqual(result, LogRecordExportResult.SUCCESS)
         self.assertEqual(self._stream.getvalue(), "")
 
-    def test_export_single_log_returns_success(self):
-        result = self._exporter.export([_make_log_record()])
+    def test_export_single_log(self):
+        result = self._exporter.export([_make_log_record("hello from test")])
         self.assertEqual(result, LogRecordExportResult.SUCCESS)
-
-    def test_export_single_log_writes_one_json_line(self):
-        self._exporter.export([_make_log_record()])
         lines = self._stream.getvalue().splitlines()
         self.assertEqual(len(lines), 1)
-        json.loads(lines[0])  # must be valid JSON
-
-    def test_export_log_body_in_output(self):
-        self._exporter.export([_make_log_record("hello from test")])
+        json.loads(lines[0])
         self.assertIn("hello from test", self._stream.getvalue())
 
-    def test_export_multiple_logs_same_resource_writes_one_line(self):
+    def test_export_multiple_logs_same_resource(self):
         logs = [
             _make_log_record("first"),
             _make_log_record("second"),
@@ -81,7 +76,7 @@ class TestFileLogExporter(unittest.TestCase):
         total_logs = sum(len(sl["logRecords"]) for sl in data["scopeLogs"])
         self.assertEqual(total_logs, 2)
 
-    def test_export_logs_different_resources_writes_multiple_lines(self):
+    def test_export_logs_different_resources(self):
         logs = [
             _make_log_record("from-a", resource_attrs={"host": "a"}),
             _make_log_record("from-b", resource_attrs={"host": "b"}),
@@ -96,52 +91,35 @@ class TestFileLogExporter(unittest.TestCase):
         exporter.export([_make_log_record()])
         mock_stream.flush.assert_called_once()
 
-    def test_custom_formatter_called(self):
+    def test_custom_formatter(self):
         formatter = Mock(return_value="formatted\n")
         exporter = FileLogExporter(self._stream, formatter=formatter)
         exporter.export([_make_log_record()])
         formatter.assert_called_once()
         self.assertIn("formatted\n", self._stream.getvalue())
 
-    def test_export_after_shutdown_returns_failure(self):
+    def test_export_after_shutdown(self):
         self._exporter.shutdown()
-        result = self._exporter.export([_make_log_record()])
+        with self.assertLogs(_LOGGER_NAME, level="WARNING"):
+            result = self._exporter.export([_make_log_record()])
         self.assertEqual(result, LogRecordExportResult.FAILURE)
-
-    def test_export_after_shutdown_logs_warning(self):
-        self._exporter.shutdown()
-        with patch.object(_log_exporter_mod._logger, "warning") as mock_warn:
-            self._exporter.export([])
-        mock_warn.assert_called_once()
-
-    def test_export_after_shutdown_writes_nothing(self):
-        self._exporter.shutdown()
-        self._exporter.export([_make_log_record()])
         self.assertEqual(self._stream.getvalue(), "")
 
-    def test_shutdown_idempotent_logs_warning(self):
+    def test_shutdown_idempotent(self):
         self._exporter.shutdown()
-        with patch.object(_log_exporter_mod._logger, "warning") as mock_warn:
+        with self.assertLogs(_LOGGER_NAME, level="WARNING"):
             self._exporter.shutdown()
-        mock_warn.assert_called_once()
 
     def test_force_flush_returns_true(self):
         self.assertTrue(self._exporter.force_flush())
 
-    def test_export_stream_error_returns_failure(self):
+    def test_export_stream_error(self):
         mock_stream = Mock()
         mock_stream.writelines.side_effect = OSError("disk full")
         exporter = FileLogExporter(mock_stream)
-        result = exporter.export([_make_log_record()])
+        with self.assertLogs(_LOGGER_NAME, level="ERROR"):
+            result = exporter.export([_make_log_record()])
         self.assertEqual(result, LogRecordExportResult.FAILURE)
-
-    def test_export_stream_error_logs_exception(self):
-        mock_stream = Mock()
-        mock_stream.writelines.side_effect = OSError("disk full")
-        exporter = FileLogExporter(mock_stream)
-        with patch.object(_log_exporter_mod._logger, "exception") as mock_exc:
-            exporter.export([_make_log_record()])
-        mock_exc.assert_called_once()
 
 
 class TestFileLogExporterIntegration(unittest.TestCase):
